+++ demo/main.qml
@@ -15,7 +15,7 @@ ApplicationWindow {
     }
 
     property var styles: [
-            "Icons", "Color Palette", "Typography"
+            "Icons", "Custom Icons", "Color Palette", "Typography"
     ]
 
     property var basicComponents: [
@@ -24,7 +24,7 @@ ApplicationWindow {
     ]
 
     property var compoundComponents: [
-            "Dialog", "Forms", "List Items", "Page Stack"
+            "Bottom Sheet", "Dialog", "Forms", "List Items", "Page Stack"
     ]
 
     initialPage: Page {
@@ -34,7 +34,15 @@ ApplicationWindow {
 
         tabs: [ "Style", "Basic Components", "Compound Components" ]
 
+        actionBar.maxActionCount: 4
+
         actions: [
+            Action {
+                iconName: "action/search"
+                name: "Search"
+                enabled: false
+            },
+
             Action {
                 iconName: "image/color_lens"
                 name: "Colors"
@@ -42,23 +50,26 @@ ApplicationWindow {
             },
 
             Action {
-                iconName: "action/search"
-                name: "Search"
+                iconName: "action/settings"
+                name: "Settings"
+                hoverAnimation: true
+            },
+
+            Action {
+                iconName: "alert/warning"
+                name: "THIS SHOULD BE HIDDEN!"
+                visible: false
             },
 
             Action {
                 iconName: "action/language"
                 name: "Language"
+                enabled: false
             },
 
             Action {
                 iconName: "action/account_circle"
-                name: "Account"
-            },
-
-            Action {
-                iconName: "action/settings"
-                name: "Settings"
+                name: "Accounts"
             }
         ]