+++ CHANGELOG.md
@@ -14,6 +14,21 @@ QML Material CHANGELOG
  * [RadioButton] Improve selection support
  * [Demo] Update radio button demo with grouping
  * [Tests] Add UI tests for the ActionBar component
+ * [IconButton] Add animation on mouse over
+ * [ListItem] Add a new expandable section header list item
+ * [Dialog] Add support for full-width content
+ * [Icons] Update the FontAwesome parsing script to pull from the website
+ * [ListItem] Correctly calculate the width of a simple menu item
+ * [ActionBar] Hide non-visible actions in the action bar
+ * [Icon] Support custom icons with an iconSource property
+ * [ActionBar] Support disabling actions in the action bar
+ * [Dialog] Correctly handle back button on Android devices
+ * [Documentation] Fix qdoc warnings
+ * [PageStack] Pop pages when the Android back button is tapped
+ * [Popup] Support preventing a popup from being dismissed
+ * [ActionBar] Control the overflow menu with Android hardware keys
+ * [ActionButton] Add support for the mini and default FAB sizes
+ * [Dialog] Expose dialog buttons via aliases
 
 
 ### Version 0.0.6 (April 16, 2015)
@@ -49,6 +64,7 @@ QML Material CHANGELOG
  * [Demo] Add the color palette to the demo
  * [ProgressBar] Port ProgressBar to inherit from QtQuick Controls
  * Minor bug fixes and code improvements
+ * [BottomSheet] Add a basic bottom sheet and an action list subclass
 
 
 ### Version 0.0.4 (February 20, 2015)